feat(refactor): refactor hooks into own file
